chore(tests): rename Foo to Child, clean things up
