+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/inherited/Child.java
@@ -13,11 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.generator;
+package io.fabric8.crd.generator.inherited;
 
-/**
- * @author <a href="claprun@redhat.com">Christophe Laprun</a>
- */
-public class BaseSpec {
+import io.fabric8.crd.generator.Base;
+import io.fabric8.kubernetes.api.model.Namespaced;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
+
+@Version("v1alpha1")
+@Group("acme.com")
+public class Child extends Base<ChildSpec, ChildStatus> implements Namespaced {
+
 
 }
