+++ kubernetes-tests/src/test/java/io/fabric8/crd/generator/inherited/ChildSpec.java
@@ -13,16 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.generator;
+package io.fabric8.crd.generator.inherited;
 
-import io.fabric8.kubernetes.api.model.Namespaced;
-import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Plural;
-import io.fabric8.kubernetes.model.annotation.Version;
-
-@Version("v1alpha1")
-@Group("acme.com")
-public class Foo extends Base<FooSpec, FooStatus> implements Namespaced {
-
+public class ChildSpec extends BaseSpec {
 
 }
